Fix bug regarding popovers in dialogs
Popovers fadeOut in order to close. The same things that trigger the
closing of a popover, trigger and the closing of the parent dialog.

So, if you trigger the closing procedure of these two, the dialog will
close and be destroyed before the popover finish the fadeOut. After
there fadeOut PopoverView tries to set value in an attribute and there
it crashes.

This could happend and in similar cases.

The solution thatwas  applied is to check in every step the _state of
the view.