ui: Handle icons indicating shared resources
Extend models of resources that can be shared with the 'shared_to_me'
and 'sharing' proxy attributes. The first is used to indicate whether a
resource is visible because it shared the project by another user, while
the second is used to return the type of the sharing (none,
shared_to_me, shared_to_project).

Also, add a small icon to show to the user if the resource is shared to
the project or shared to the user.

Signed-off-by: Christos Stavrakakis <stavr.chris@gmail.com>